Move extract_if iterator into its own module
Move BtreeExtractIf out of btree_iters.rs and into tree_store/extract_if.rs without changing behavior. Keep btree_iters focused on range iteration and re-export BtreeExtractIf from tree_store.
